src/app/fizzbuzz/fizzbuzz.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { FizzbuzzComponent } from './fizzbuzz.component';

describe('FizzbuzzComponent', () => {
  let component: FizzbuzzComponent;
  let fixture: ComponentFixture<FizzbuzzComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ FizzbuzzComponent ]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(FizzbuzzComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should return valid fizzbuzz value for 10', () => {
    expect(component.getMappedEntries(20)[9]).toEqual('Buzz');
  });

  it('should return valid fizzbuzz value for 15', () => {
    expect(component.getMappedEntries(20)[14]).toEqual('FizzBuzz');
  });
});
